+++ alinkjs/packages/assetlink-default-plugins/plugins/AssetPageTitleWithLocationsWidgetDecorator.alink.vue
@@ -12,50 +12,10 @@ const props = defineProps({
 
 const assetLink = inject('assetLink');
 
-const loadingLocations = ref(false);
 const currentLocations = ref(null);
 
-const resolveCurrentLocation = async () => {
-  loadingLocations.value = true;
-
-  const logTypes = (await assetLink.getLogTypes()).map(t => t.attributes.drupal_internal__id);
-
-  const populateLocationsFromLatestMovementLogs = async (entitySource, entitySourceCache) => {
-
-    const results = await entitySource.query(q => logTypes.map(logType => {
-      return q.findRecords(`log--${logType}`)
-        .filter({ attribute: 'is_movement', op: 'equal', value: true })
-        .filter({ attribute: 'status', op: 'equal', value: 'done' })
-        .filter({ attribute: 'timestamp', op: '<=', value: currentEpochSecond() })
-        .filter({
-          relation: 'asset.id',
-          op: 'some',
-          records: [{ type: props.asset.type, id: props.asset.id }]
-        })
-        .sort('-timestamp')
-        .page({ offset: 0, limit: 1 });
-    }), {
-      sources: {
-        remote: {
-          include: ['location']
-        }
-      }
-    });
-
-    const logs = results.flatMap(l => l);
-
-    const latestLog = logs.length ? logs.reduce((logA, logB) => parseJSONDate(logA.attributes.timestamp) > parseJSONDate(logB.attributes.timestamp) ? logA : logB) : null;
-
-    if (latestLog) {
-      currentLocations.value = await entitySourceCache.query(q => q.findRelatedRecords(latestLog, 'location'))
-    }
-  }
-
-  await populateLocationsFromLatestMovementLogs(assetLink.entitySource.cache, assetLink.entitySource.cache);
-
-  await populateLocationsFromLatestMovementLogs(assetLink.entitySource, assetLink.entitySource.cache);
-
-  loadingLocations.value = false;
+const resolveCurrentLocation = () => {
+  currentLocations.value = assetLink.entitySource.cache.query(q => q.findRelatedRecords({ type: props.asset.type, id: props.asset.id }, 'location'))
 };
 
 const onAssetLogsChanged = ({ assetType, assetId }) => {